refactor(agent): organize action runtime modules
